authorization

+++ backend/db/repository/blog.py
@@ -25,8 +25,12 @@ def list_blogs(db: Session):
 
 def update_blog_by_id(id: int, blog:UpdateBlog, db:Session, author_id: int=1):
     blog_in_db = db.query(Blog).filter(Blog.id==id).first()
+
     if not blog_in_db:
-        return
+        return {"error": f"blod with id {id} does not exists"}
+    if not blog_in_db.author_id==author_id:
+        return {"error": f"Only the author can modify the blog"}
+    
     blog_in_db.title = blog.title
     blog_in_db.content = blog.content
     db.add(blog_in_db)
@@ -35,9 +39,13 @@ def update_blog_by_id(id: int, blog:UpdateBlog, db:Session, author_id: int=1):
 
 def delete_blog_by_id(id: int, db: Session, author_id: int):
     blog_in_db = db.query(Blog).filter(Blog.id==id)
+    print(id, author_id)
     if not blog_in_db.first():
         return{"error": f"could not find blog with id {id}"}
-    
+    if not blog_in_db.first().author_id==author_id:
+        return {
+            "error": "Only author can delete a blog"
+        }
     blog_in_db.delete()
     db.commit()
     return {"msg": f"Deleted blog with id {id}"}
